pyproject.toml
[project]
name = "jax_alqp"
version = "0.1"
requires-python = ">= 3.12"
dependencies = [
  "absl-py",
  "jax[cpu]; platform_system != \"Linux\"",
  "jax[cuda12]; platform_system == \"Linux\"",
  "numpy",
  "primal_dual_ilqr @ git+https://github.com/joaospinto/primal_dual_ilqr",
]
authors = [
  {name = "Joao Sousa-Pinto", email = "joaospinto@gmail.com"},
]
description = "JAX Augmented Lagrangian QP solver"
license = "MIT"

[tool.black]
line-length = 80

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["jax_alqp"]
